Added DTO'S

+++ rent-a-car/rent-a-car/src/main/java/com/example/rentacar/repository/BrandRepository.java
@@ -1,14 +1,12 @@
 package com.example.rentacar.repository;
 
 
-import com.example.rentacar.entites.Brand;
+import com.example.rentacar.entity.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 //CRUD Operations
 //Repositry anotasyonunu eklemek zorunda degiliz.
-public interface BrandRepository extends JpaRepository<Brand,Integer> {
+public interface BrandRepository extends JpaRepository<Brand,Long> {
     //Custom queries
